backend/requirements.txt
# ── Core ─────────────────────────────────────────────────────────────────────
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv

# ── Database ──────────────────────────────────────────────────────────────────
motor
pymongo

# ── Data & Feature Engineering ────────────────────────────────────────────────
pandas
numpy
pandas-ta
scipy

# ── ML ────────────────────────────────────────────────────────────────────────
scikit-learn
xgboost
joblib

# ── Deep Learning ─────────────────────────────────────────────────────────────
torch --index-url https://download.pytorch.org/whl/cpu

# ── OANDA / HTTP ─────────────────────────────────────────────────────────────
requests
httpx

# ── LLM ──────────────────────────────────────────────────────────────────────
google-generativeai
openai

# ── MLflow ───────────────────────────────────────────────────────────────────
mlflow

# ── Websocket ─────────────────────────────────────────────────────────────────
websockets
